Fold PALWindows into concept-organized PAL files
A platform-named file is the growth pattern the PAL design rule exists
to prevent. The FILETIME/Attributes conversions move next to their POSIX
counterparts in PALFileDescriptor.swift, and the Win32-error-to-errno
table moves into PALErrnoMapping.swift, which now owns the whole error
story: foreign error namespaces into errno, errno into WASI codes.
